The 'Recover last session' feature is fully implemented now.
* vitables/vtapp.py
getSessionFilesNodes method has changed. Code for taking care of leaves with
views has been added.
slotNodeOpen method has been debugged. Now it works fine when the last session
is being recovered and a node has to be open.
A lot of stuff that was commented out has been removed because it was
definitely unneeded.