Fixed an annoyance in the creation of PDF documentation.
The combination fop + SVG can be tricky because SVG rendering relies on the
AWT, which in turn relies on the native windows system, a running X server on
Unixes. As a consequence, using a SVG file for generating the title page of
the User's Guide in PDF format can result in fop errors if the shell where the
installation process is running is not connected to the X server. This errors
crash the installation process. Although the fix is simple (connect the shell
to the X server via xhost + from a different shell and export DISPLAY) it is
very annoying.

The problem is avoided by using a PNG file for generating the PDF version of
the User's Guide.

* MANIFEST.in
File has been updated. Unneeded SVG files are excluded from the package. The
XSL custom layer for generating the documentation has beeen added.

* doc/usersguide.xml
File has beeen updated.

* doc/images/title_page_plain.png
File has been added.