Fixed a minor bug in the code for opening files via drag and drop.
The following problems have been fixed:

- files opening via D&D did'nt work well on Linux platforms
- the Users Guide was not being included in the ViTables tarball

* vitables/h5db/dbsTreeModel.py
dropMimeData method has been refactored and debugged.

* MANIFEST.in
The Users Guide in PDF and HTML format is now included in the tarball

* setup.py
BuildDocbook.finalize_options method changed. Now it will not abort the user's
installation if xsltproc/fop are not found.
BuildDocbook.run method debugged. Now it gracefully returns if documentation
cannot be built. It also adds the css stylesheet to the output directory.

* mswindows/Makefile_windows
Now it adds the css stylesheet to the created output directory.