Several methods renamed to shorter names. Fixed a small bug.
Several methods have been renamed in the vtapp.py module. New names
are shorter
than the old ones and improve code readability.

Improved the disambiguation in the Ctrl+C keyboard shortcut. It is
used by both
the logger and the databases tree view widgets.

Fixed a small bug when dealing with the QAction.triggered signal. It
is emitted
whit a boolean argument and we cannot use None any more for dealing
with this
argument in the slots connected to this signal.