Fixed some minor bugs regarding plugins.
When a plugins path is removed from the list of searchable paths then
the plugins living under this path are removed from the list of
enabled/disabled/all plugins.

The plugins_menu entry of the Plugins menu is not forced to be the
first entry now.

The items of the QListViews of the Plugins page (Preferences dialog) are
not editable now.