Bug 338 has been fixed.
This subtle bug  was hitting not only to group creation
but to
every node editing if a editing dialog was involved. So the following
operations were hit by this bug: launching the Properties dialog,
deleting a
node, renaming (not in-place) a node, creating a group, pasting a node
(with
possible overwriting) and moving a node (with possible overwriting.