A selected cell rendering bug has been fixed.
A custom delegate is used in order to render the LeafView items. The
delegate has an overriden paint() method which is able to distinguish if
the rendered selected cell matches the displayed buffer. If it matches
then the selected cell is rendered in the standard way. If not then it
is rendered as a regular cell.