The shebang/encoding lines of all modules have been updated.
The change ensures consistency with the scripts/vitables launcher
script. Also it seems more secure not to use the generic python command
as it could invoke Python3 if present (which is currently not desired): as
far as I know the generic python command points to the *last* python
interpreter installed on a system so it can call python2 or python3
depending on the order of installation.

In theory this change doesn't affect to Windows platforms.