Fix bug in PluginsLoader.load function.
If a plugin was enabled but later removed from disk then the program
would crash. KeyError exception occured in

plugin = self.all_plugns[UID]

finally block tried to use  finding_failed variable which was not
defined because it was initialized after the above line. This
was causing another exception that was not caught.

Added logger and some log messages.

Added handler for KeyError exception.