Track data files without requiring setuptools-git
As discussed in uvemas/ViTables#33, the data files for the development
branch are not tracked unless the user has setuptools-git installed.
However, it is not guaranteed that all users will have this requirement
as not all end users are necessarily developers or have the proper git
tools installed.  This patch adds a function to the setup.py script to
traverse the packages and locate data files in all subdirectories that
are not subpackages.  Data files are assumed to be all files that are
not Python files (.py and .pyc).