Convert lists to arrays in calculator
The tool tip on the calculator states that lists are valid objects to
return from the expression.  However, in practice the following  error
is raised

    objects of type ``list`` are not supported in this context, sorry;
    supported objects are: NumPy array, record, or scalar

Digging around, this error is due to the native Python types being
removed as a valid flavor on line 49 of `vitables/vttables/buffer.py`.
To correct this while maintaining that all returned leaves are NumPy
arrays (per the comments above line 49), this patch explicitly checks if
the result returned by `vitables.calculator.evaluator.evaluate` on line
347 of `vitables/calculatlr/calculator.py` is a list or tuple.  If so,
it is converted to a NumPy array; otherwise, it is assumed scalar and
converted to a NumPy array of shape `(1,)`.