Gracefully handle empty fields
If the statements, expression, or destination is empty when the
expression is saved, this patch makes sure the empty string is loaded
when the expression is chosen.  This quites errors about QPuNullVarient
being an invalid argument.

By allowing null values, the user may save a partially complete
expression to return to it later.