Deal with an empty list when remove is clicked
If the list of stored expressions is empty, don't bother trying to
update the selected item which triggers an index out of bounds error.
This patch ensures that an item exists in the selected indexes list
before trying to index it.  The first attempt checked
`self.saved_list.count()` to ensure at leas one item was in the list.
However, a race condition exists when clicking the "Remove" button for
the last item in the list.  Both `on_remove_button_clicked` and
`on_saved_list_itemSelectionChanged` are dispatched when the button is
clicked meaning `count()` is not guaranteed to give the correct number
of items in `self.saved_list`.

To reproduce the race condition, place print statements at the start and
end of `on_remove_button_clicked` and
`on_saved_list_itemSelectionChanged`.  Then store an expression and
remove it.  With luck, the messages will intermingle.