fix(dbtree): SINGLE-click select nodes, DBLE-click open/activate views
## Change of GUI behavior
- Now, clicking on a leaf-node, selects only the node - does not open
it.
- Double-clicking, either opens it in a subWindow, or 'activates' it, if
already open.

## Rational:
So far, signle-clicking immediately activated that sub window.
But then node was loosing focus, surprising the user; old code was
trying to regain focus on the node, but it failed to work (at least in
Qt5)...
That disrupted e.g. the selection of multiple leaf-nodes, if any one of
them had sub-windows open.